requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.39.0
Django==3.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
djangorestframework==3.15.0
gunicorn==21.2.0
idna==3.6
packaging==24.0
pillow==10.2.0
psycopg2==2.9.9
pytz==2024.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
